refactor: Refactored doctr.models
Split Predictor objects into framework-specific implementations
