+++ doctr/models/__init__.py
@@ -1,8 +1,4 @@
-from .preprocessor import *
-from .core import *
 from . import artefacts
-from . import utils
-from ._utils import *
 from .backbones import *
 from .detection import *
 from .recognition import *
