+++ doctr/models/detection/__init__.py
@@ -1,4 +1,3 @@
-from .core import *
 from .differentiable_binarization import *
 from .linknet import *
 from .zoo import *
