+++ doctr/models/detection/core.py
@@ -5,23 +5,14 @@
 
 import numpy as np
 import cv2
-from typing import List, Any, Optional, Dict, Tuple
+from typing import List, Tuple
 
 from doctr.utils.repr import NestedObject
 from doctr.utils.geometry import rotate_image
 from .._utils import get_bitmap_angle
-from .. import PreProcessor
 
 
-__all__ = ['DetectionModel', 'DetectionPostProcessor', 'DetectionPredictor']
-
-
-class DetectionModel(NestedObject):
-    """Implements abstract DetectionModel class"""
-
-    def __init__(self, cfg: Optional[Dict[str, Any]] = None) -> None:
-        super().__init__()
-        self.cfg = cfg
+__all__ = ['DetectionPostProcessor']
 
 
 class DetectionPostProcessor(NestedObject):
@@ -115,40 +106,3 @@ class DetectionPostProcessor(NestedObject):
             boxes_batch.append(boxes)
 
         return boxes_batch, angles_batch
-
-
-class DetectionPredictor(NestedObject):
-    """Implements an object able to localize text elements in a document
-
-    Args:
-        pre_processor: transform inputs for easier batched model inference
-        model: core detection architecture
-    """
-
-    _children_names: List[str] = ['pre_processor', 'model']
-
-    def __init__(
-        self,
-        pre_processor: PreProcessor,
-        model: DetectionModel,
-    ) -> None:
-
-        self.pre_processor = pre_processor
-        self.model = model
-
-    def __call__(
-        self,
-        pages: List[np.ndarray],
-        **kwargs: Any,
-    ) -> List[np.ndarray]:
-
-        # Dimension check
-        if any(page.ndim != 3 for page in pages):
-            raise ValueError("incorrect input shape: all pages are expected to be multi-channel 2D images.")
-
-        processed_batches = self.pre_processor(pages)
-        predicted_batches = [
-            self.model(batch, return_boxes=True, **kwargs)['preds']  # type:ignore[operator]
-            for batch in processed_batches
-        ]
-        return [pred for batch in predicted_batches for pred in zip(*batch)]
