+++ doctr/models/recognition/__init__.py
@@ -1,7 +1,4 @@
-from .core import *
 from .crnn import *
 from .master import *
 from .sar import *
 from .zoo import *
-
-del utils  # type: ignore[name-defined]
