+++ doctr/models/zoo.py
@@ -4,7 +4,7 @@
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
 
 from typing import Any
-from .core import OCRPredictor
+from .predictor import OCRPredictor
 from .detection.zoo import detection_predictor
 from .recognition.zoo import recognition_predictor
